[ai-assisted] test(attachment): address PR #166 review feedback
Validation:
- ./gradlew -p /tmp/studio-api-issue-161 :studio-application-modules:attachment-service:test --tests 'studio.one.application.web.controller.AttachmentAccessSupportTest' --tests 'studio.one.application.web.controller.AttachmentControllerTest' --tests 'studio.one.application.web.controller.AttachmentMgmtControllerAuthorizationTest' --tests 'studio.one.application.web.controller.AttachmentWebSupportTest' --tests 'studio.one.application.web.controller.MeAttachmentControllerTest'
- ./gradlew -p /tmp/studio-api-issue-161 :studio-application-modules:attachment-service:compileJava
