+++ CHANGELOG.md
@@ -9,8 +9,8 @@
 - `attachment-service`의 접근 제어 helper를 `AttachmentAccessSupport`로 분리해 principal 조회, 관리자 판별, owner 접근 검사를 컨트롤러에서 공통으로 사용하도록 정리했다.
 
 ### 검증
-- `./gradlew -p /tmp/studio-api-issue-161 :studio-application-modules:attachment-service:test --tests 'studio.one.application.web.controller.AttachmentControllerTest' --tests 'studio.one.application.web.controller.AttachmentMgmtControllerAuthorizationTest' --tests 'studio.one.application.web.controller.AttachmentWebSupportTest' --tests 'studio.one.application.web.controller.MeAttachmentControllerTest'`
-- `./gradlew -p /tmp/studio-api-issue-161 :studio-application-modules:attachment-service:compileJava`
+- `./gradlew :studio-application-modules:attachment-service:test --tests 'studio.one.application.web.controller.AttachmentAccessSupportTest' --tests 'studio.one.application.web.controller.AttachmentControllerTest' --tests 'studio.one.application.web.controller.AttachmentMgmtControllerAuthorizationTest' --tests 'studio.one.application.web.controller.AttachmentWebSupportTest' --tests 'studio.one.application.web.controller.MeAttachmentControllerTest'`
+- `./gradlew :studio-application-modules:attachment-service:compileJava`
 
 ## 2026-03-30
