+++ studio-application-modules/attachment-service/src/test/java/studio/one/application/web/controller/AttachmentAccessSupportTest.java
@@ -1,0 +1,37 @@
+package studio.one.application.web.controller;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import java.util.Set;
+
+import org.junit.jupiter.api.Test;
+
+import studio.one.platform.identity.ApplicationPrincipal;
+
+class AttachmentAccessSupportTest {
+
+    @Test
+    void isAdminAcceptsLegacyAndSpringSecurityRoleNames() {
+        assertTrue(AttachmentAccessSupport.isAdmin(principal("ADMIN")));
+        assertTrue(AttachmentAccessSupport.isAdmin(principal("ROLE_ADMIN")));
+    }
+
+    private ApplicationPrincipal principal(String role) {
+        return new ApplicationPrincipal() {
+            @Override
+            public Long getUserId() {
+                return 1L;
+            }
+
+            @Override
+            public String getUsername() {
+                return "admin";
+            }
+
+            @Override
+            public Set<String> getRoles() {
+                return Set.of(role);
+            }
+        };
+    }
+}
