+++ studio-application-modules/attachment-service/src/test/java/studio/one/application/web/controller/AttachmentMgmtControllerAuthorizationTest.java
@@ -77,7 +77,7 @@ class AttachmentMgmtControllerAuthorizationTest {
         AttachmentMgmtController controller = controller();
 
         when(principalResolverProvider.getIfAvailable()).thenReturn(principalResolver);
-        when(principalResolver.currentOrNull()).thenReturn(principal(1L, "ROLE_ADMIN"));
+        when(principalResolver.currentOrNull()).thenReturn(principal(1L, "ADMIN"));
         when(attachmentService.getAttachments(12, 34L)).thenReturn(Collections.emptyList());
 
         controller.listByObject(12, 34L);
